+++ corehq/form_processor/utils/__init__.py
@@ -1,0 +1,15 @@
+from .general import (
+    should_use_sql_backend,
+)
+
+from .xform import (
+    extract_meta_instance_id,
+    new_xform,
+    convert_xform_to_json,
+    adjust_datetimes,
+    acquire_lock_for_xform,
+)
+
+from .metadata import (
+    clean_metadata,
+)
